# contracts/openapi/pie-control-plane-v1.yaml
openapi: 3.1.2
info:
  title: Pie Control Plane API
  version: 0.1.0
  description: Initial executable contract for instance discovery, session, organizations, projects, and work items.
servers:
  - url: https://api.pielab.ai
security:
  - PieOidc: []
...
paths:
...
  /v1/organizations/{organizationId}/meetings/{meetingId}/waiting-room:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
      - name: meetingId
        in: path
        required: true
        schema:
          $ref: ../schemas/common/opaque-id.v1.schema.json
    post:
      tags: [Meetings]
      operationId: requestMeetingAdmission
      description: Requests admission or returns the caller's current waiting-room status.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Current participant admission state.
          content:
            application/json:
              schema:
                $ref: ../schemas/resources/meeting-participant.v1.schema.json
        default:
          $ref: '#/components/responses/Problem'
  /v1/organizations/{organizationId}/meetings/{meetingId}/media-diagnostics:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
      - name: meetingId
        in: path
        required: true
        schema:
          $ref: ../schemas/common/opaque-id.v1.schema.json
    get:
      tags: [Meetings]
      operationId: diagnoseMeetingMedia
      description: Checks control-plane and LiveKit endpoint reachability without joining the room.
      responses:
        '200':
          description: Current meeting media diagnostic.
          content:
            application/json:
              schema:
                $ref: ../schemas/resources/meeting-media-diagnostics.v1.schema.json
        default:
          $ref: '#/components/responses/Problem'
  /v1/organizations/{organizationId}/meetings/{meetingId}/media-token:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
      - name: meetingId
        in: path
        required: true
        schema:
          $ref: ../schemas/common/opaque-id.v1.schema.json
    post:
      tags: [Meetings]
      operationId: issueMeetingMediaToken
      description: Issues a short-lived room token only after server-side admission and role checks.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Capability-scoped LiveKit token.
          content:
            application/json:
              schema:
                $ref: ../schemas/resources/meeting-media-token.v1.schema.json
        default:
          $ref: '#/components/responses/Problem'
...
  /v1/organizations/{organizationId}/meetings/{meetingId}/governance:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
      - name: meetingId
        in: path
        required: true
        schema:
          $ref: ../schemas/common/opaque-id.v1.schema.json
    get:
      tags: [Meetings]
      operationId: getMeetingGovernance
      responses:
        '200':
          description: Meeting capture and retention policy.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: ../schemas/resources/meeting-governance.v1.schema.json
        default:
          $ref: '#/components/responses/Problem'
    patch:
      tags: [Meetings]
      operationId: updateMeetingGovernance
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: ../schemas/resources/meeting-governance-update.v1.schema.json
      responses:
        '200':
          description: Updated meeting governance policy.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: ../schemas/resources/meeting-governance.v1.schema.json
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    PieOidc:
      type: openIdConnect
      openIdConnectUrl: https://auth.pielab.ai/realms/pie/.well-known/openid-configuration
      description: The actual issuer is discovered from /.well-known/pie for each instance.
  parameters:
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema:
        $ref: ../schemas/common/opaque-id.v1.schema.json
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
        pattern: '^"[^" ]+"$'
  headers:
    ETag:
      description: Strong entity tag for the returned resource version.
      required: true
      schema:
        type: string
        pattern: '^"[^" ]+"$'
...
  responses:
    Problem:
      description: RFC 9457 problem with a stable Pie error code.
      content:
        application/problem+json:
          schema:
            $ref: ../schemas/common/problem-details.v1.schema.json